+++ packages/sdk/src/prepareOperation.ts
@@ -28,6 +28,7 @@ import {
   groupByArr,
 } from "./utils";
 import { SparseMerkleProver } from "./SparseMerkleProver";
+import { Logger } from "winston";
 
 export const __private = {
   gatherNotes,
@@ -37,6 +38,7 @@ export interface PrepareOperationDeps {
   db: NocturneDB;
   viewer: NocturneViewer;
   merkle: SparseMerkleProver;
+  logger?: Logger;
 }
 
 export async function prepareOperation(
@@ -51,6 +53,7 @@ export async function prepareOperation(
   let joinSplits: PreSignJoinSplit[] = [];
   const usedMerkleIndices = new Set<number>();
   for (const joinSplitRequest of joinSplitRequests) {
+    console.log("preparing joinSplits for request: ", joinSplitRequest);
     const newJoinSplits = await prepareJoinSplits(
       deps,
       joinSplitRequest,
@@ -58,8 +61,13 @@ export async function prepareOperation(
     );
 
     newJoinSplits.forEach((js) => {
-      usedMerkleIndices.add(js.oldNoteA.merkleIndex);
-      usedMerkleIndices.add(js.oldNoteB.merkleIndex);
+      // If note value == 0, its just a dummy and we don't want to count in used merkle indices
+      if (js.oldNoteA.value !== 0n) {
+        usedMerkleIndices.add(js.oldNoteA.merkleIndex);
+      }
+      if (js.oldNoteB.value !== 0n) {
+        usedMerkleIndices.add(js.oldNoteB.merkleIndex);
+      }
     });
     joinSplits.push(...newJoinSplits);
   }
@@ -89,7 +97,7 @@ export async function prepareOperation(
 }
 
 async function prepareJoinSplits(
-  { db, viewer, merkle }: PrepareOperationDeps,
+  { db, viewer, merkle, logger }: PrepareOperationDeps,
   joinSplitRequest: JoinSplitRequest,
   alreadyUsedNoteMerkleIndices: Set<number> = new Set()
 ): Promise<PreSignJoinSplit[]> {
@@ -97,7 +105,8 @@ async function prepareJoinSplits(
     db,
     getJoinSplitRequestTotalValue(joinSplitRequest),
     joinSplitRequest.asset,
-    alreadyUsedNoteMerkleIndices
+    alreadyUsedNoteMerkleIndices,
+    logger
   );
 
   const unwrapAmount = joinSplitRequest.unwrapValue;
@@ -122,8 +131,11 @@ async function gatherNotes(
   db: NocturneDB,
   requestedAmount: bigint,
   asset: Asset,
-  noteMerkleIndicesToIgnore: Set<number> = new Set()
+  noteMerkleIndicesToIgnore: Set<number> = new Set(),
+  logger?: Logger
 ): Promise<IncludedNote[]> {
+  console.log("indices to ignore", noteMerkleIndicesToIgnore);
+
   // check that the user has enough notes to cover the request
   const notes = (await db.getNotesForAsset(asset)).filter(
     (n) => !noteMerkleIndicesToIgnore.has(n.merkleIndex)
@@ -178,7 +190,11 @@ async function gatherNotes(
     subseqIndex--;
   }
 
-  console.log("notes to use: ", notesToUse);
+  logger &&
+    logger.debug(
+      `gathered notes to satisfy request for ${requestedAmount} of assest ${asset.assetAddr}`,
+      { notesToUse, requestedAmount, asset }
+    );
   return notesToUse;
 }