+++ packages/offchain-utils/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 
+- move redis out to remove peer-dependency on `ioredis`
 - add `opentelemetry` default instrumentation + metric name formatting util
 - add `HealthCheckResponse`
 - move redis, ajv, and actor related utils to this pkg so bundler and screener can share code